Installs kwbdi vim plugin, better buffer closer
Using \bd will close the current buffer.
The window layout will be kept in all circumstances.
If there is an alternative buffer that will be shown.
If there is no alternative buffer then the previous buffer will be
shown.
If there is no alternative nor previous buffer then an empty buffer will
be shown.