Remove old layout
This basically destroys this project. Going to recreate this as a set of
ansible roles, that way dependencies between them can be managed and the
required operations are performed in a more consistent way.

Aim is:

 * Retain full existing functionality
 * Add dependency information for different roles
 * Break down monolithic parts into smaller units

Eventually separate installs (server, local, location specific etc) can
be created which can apply the ansible roles to the machine that they
are run on. Then finally the dotfiles will be installable with some
command like:

curl http://github.com/matthewfranglen/dotfiles/install.sh | bash

This is also a chance to clear up the correct order for install /
install-dependencies / make once and for all.