Fix use of copy command
After reviewing the source of the copy command I have found that it
treats trailing slashes as an indication of handling more in line with
what is desired (not creating inner directories when copying
directories). Hopefully this will fix the folder-within-folder issue and
then thoughts about provisioning remote hosts can be had.